+++ lib/src/widgets/reactive_form_field.dart
@@ -14,12 +14,14 @@ typedef ReactiveFormFieldBuilder<T, K> = Widget Function(
     ReactiveFormFieldState<T, K> field);
 
 /// Signature for customize when to show errors in a widget.
-typedef ShowErrorsFunction = bool Function(AbstractControl<dynamic> control);
+typedef ShowErrorsFunction<T> = bool Function(FormControl<T> control);
 
 /// Signature of the function that returns the [Map] that store custom
 /// validation messages for each error.
-typedef ValidationMessagesFunction<T> = Map<String, String> Function(
-    FormControl<T> control);
+typedef ValidationMessageFunction = String Function(Object error);
+
+/// Signature of a callback that provides the control as argument.
+typedef ReactiveFormFieldCallback<T> = void Function(FormControl<T> control);
 
 /// A single reactive form field.
 ///
@@ -39,15 +41,15 @@ class ReactiveFormField<ModelDataType, ViewDataType> extends StatefulWidget {
   /// The control that is bound to this widget.
   final FormControl<ModelDataType>? formControl;
 
-  /// A function that returns the [Map] that store custom validation messages
+  /// A function that returns the [Map] that stores custom validation messages
   /// for each error.
-  final ValidationMessagesFunction<ModelDataType>? validationMessages;
+  final Map<String, ValidationMessageFunction>? validationMessages;
 
   /// Gets the widget control value accessor
   final ControlValueAccessor<ModelDataType, ViewDataType>? valueAccessor;
 
   /// Gets the callback that define when to show errors in UI.
-  final ShowErrorsFunction? showErrors;
+  final ShowErrorsFunction<ModelDataType>? showErrors;
 
   /// Creates an instance of the [ReactiveFormField].
   ///
@@ -99,28 +101,24 @@ class ReactiveFormFieldState<ModelDataType, ViewDataType>
   /// If the control has several errors, then the first error is selected
   /// for visualizing in UI.
   String? get errorText {
-    if (_showErrors()) {
-      final validationMessages = _getValidationMessages(control);
-      return validationMessages.containsKey(control.errors.keys.first)
-          ? validationMessages[control.errors.keys.first]
-          : control.errors.keys.first;
+    if (control.hasErrors && _showErrors) {
+      final errorKey = control.errors.keys.first;
+      final validationMessage = _findValidationMessage(errorKey);
+
+      return validationMessage != null
+          ? validationMessage(control.getError(errorKey)!)
+          : errorKey;
     }
 
     return null;
   }
 
-  bool _showErrors() {
+  bool get _showErrors {
     if (widget.showErrors != null) {
       return widget.showErrors!(control);
     }
 
     return control.invalid && touched;
-  }
-
-  Map<String, String> _getValidationMessages(FormControl<dynamic> control) {
-    return widget.validationMessages != null
-        ? widget.validationMessages!(this.control)
-        : <String, String>{};
   }
 
   @override
@@ -255,5 +253,15 @@ class ReactiveFormFieldState<ModelDataType, ViewDataType>
   @override
   Widget build(BuildContext context) {
     return widget._builder(this);
+  }
+
+  ValidationMessageFunction? _findValidationMessage(String errorKey) {
+    if (widget.validationMessages != null &&
+        widget.validationMessages!.containsKey(errorKey)) {
+      return widget.validationMessages![errorKey];
+    } else {
+      final formConfig = ReactiveFormConfig.of(context);
+      return formConfig?.validationMessages[errorKey];
+    }
   }
 }